# api/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Paperless REST API
  version: v1.0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /users/register:
    post:
      tags:
        - user-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /users/login:
    post:
      tags:
        - user-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    AuthDTO:
      type: object
      properties:
        username:
          type: string
          maxLength: 20
          minLength: 3
        password:
          type: string
          maxLength: 2147483647
          minLength: 8
      required:
        - password
        - username
...